Fix code duplication by moving function
