+++ data_analysis/electricity_plans/electricity_plans_analysis.py
@@ -24,11 +24,11 @@ import logging
 from data_analysis.plan_choice_helpers.data_loading import load_tariff_data
 from data_analysis.plan_choice_helpers.electricity_plan_helpers import (
     filter_electricity_plans,
-    plot_subset,
 )
 from data_analysis.plan_choice_helpers.general_helpers import (
     clear_output_dir,
     generate_pdf_from_png,
+    plot_subset,
 )
 from data_analysis.postcode_lookup_tables.geo_utils import load_and_transform_shapefile
